servicio trabajadores back y tabla de agregar trabajadores en el front

+++ geos/src/main/java/com/geos/repository/IUsuarioRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.geos.entity.Usuarios;
 
-public interface IUsuarioRepository extends CrudRepository<Usuarios, Long> {
+public interface IUsuarioRepository extends CrudRepository<Usuarios, Integer> {
 
 	Usuarios findByRut(String rut);
